+++ package.json
@@ -21,7 +21,6 @@
     "cookie-parser": "^1.4.3",
     "cos-nodejs-sdk-v5": "^1.1.0",
     "express": "^4.15.2",
-    "jimp": "^0.2.28",
     "mongoose": "^4.9.8",
     "multer": "^1.3.0"
   }
