+++ view/edit-ajax.js
@@ -2,8 +2,6 @@ var artist=new Object();
 $.get("http://xx.com/v1/user/info",function(user,result){
 	artist=user;
 	if(result=="success"){
-		
+		initilize_artist(user.data)
 	}
 }
-
-
